# test-project/care-clarity-export/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Care Clarity API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Source language
 */
export type TranslateRequestFrom =
  (typeof TranslateRequestFrom)[keyof typeof TranslateRequestFrom];

export const TranslateRequestFrom = {
  es: "es",
  en: "en",
} as const;

/**
 * Target language
 */
export type TranslateRequestTo =
  (typeof TranslateRequestTo)[keyof typeof TranslateRequestTo];

export const TranslateRequestTo = {
  es: "es",
  en: "en",
} as const;

export interface TranslateRequest {
  /** The text to translate */
  text: string;
  /** Source language */
  from: TranslateRequestFrom;
  /** Target language */
  to: TranslateRequestTo;
}

export interface TranslateResponse {
  /** The translated text */
  translatedText: string;
}

export interface ErrorResponse {
  error: string;
}