# test-project/care-clarity-export/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Care Clarity API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Translates text between English and Spanish using MyMemory API
 * @summary Translate text
 */
export const TranslateBody = zod.object({
  text: zod.string().describe("The text to translate"),
  from: zod.enum(["es", "en"]).describe("Source language"),
  to: zod.enum(["es", "en"]).describe("Target language"),
});

export const TranslateResponse = zod.object({
  translatedText: zod.string().describe("The translated text"),
});