# test-project/care-clarity-export/lib/api-zod/src/generated/types/translateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Care Clarity API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TranslateRequestFrom } from "./translateRequestFrom";
import type { TranslateRequestTo } from "./translateRequestTo";

export interface TranslateRequest {
  /** The text to translate */
  text: string;
  /** Source language */
  from: TranslateRequestFrom;
  /** Target language */
  to: TranslateRequestTo;
}