# test-project/care-clarity-export/lib/api-zod/src/generated/types/translateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Care Clarity API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranslateResponse {
  /** The translated text */
  translatedText: string;
}